v 1.0.0
The first major refactor! Split up the code and moved a lot of stuff to helper files. A lot of the logic is still in the main file, since the added abstraction was more of a pain than a help if I'm being honest.

Major changes include:
- Config file to help make it easy to set preferences
- Queue creation is now restricted to admins only to avoid abuse
- Limited intentions to prevent possible damage from misuse
- Option to give previous party members queue priority after a party is disbanded

There's also some minor bugfixes, most notably that requeued players will now be properly requeued if a pop occurs during the requeue process.